+++ zcash_script/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Bumped `bindgen` dependency from `>= 0.69.5` to `>= 0.71`.
+- The `zcash_script::opcode::PossiblyBad::parse` fn is now `pub`.
 
 ## [0.4.2], [0.5.1] - 2025-10-24