+++ zcash_script/CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+### Changed
+
+- Bumped `bindgen` dependency from `>= 0.69.5` to `>= 0.71`.
+
 ## [0.4.2], [0.5.1] - 2025-10-24
 
 ### Fixed