+++ zcash_script/CHANGELOG.md
@@ -8,7 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+<<<<<<< HEAD
 ## [0.4.2] - 2025-10-24
+=======
+## Added
+
+- `zcash_script::script::Code::is_unspendable()`
+
+## [0.4.2], [0.5.1] - 2025-10-24
+>>>>>>> 4c631e1 (make Code::is_unspendable() public (#280))
 
 ### Fixed
 - Fixed `no-std` support which was broken by [#256].